requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LLM İletişimi
requests==2.31.0

# Vektör Veritabanı ve Embeddings
faiss-cpu==1.7.4
sentence-transformers==2.3.1

# Doküman İşleme
pypdf==4.0.1
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0

# Testing
pytest==7.4.4
httpx==0.26.0

# Dependencies
numpy==1.24.3
